closed for inactivity for over a year
Lisa <lisa.jones@owasp.com>
11:46 AM (14 minutes ago)

 1
to basta.aditya2019, Kunal, Nithyanandam, Shraddha, Sparsh, OWASP-Chapter-Committee

Leaders,

I am unable to locate any recent chapter meeting activity. The chapter is not using the free Meetup Group provided, and the last post on the OWASP chapter page is for an event that charged a registration fee without approval from the OWASP  Foundation in October of 2021. WASPCON 2021

The Meetup group members will be emailed today that the group is closed for inactivity. The leaders have been removed from the OWASP chapter page, and the chapter will be closed for inactivity for over a year. The leaders will receive an automated email about the closing of their OWASP account, or they can join and become a member to join. 

All other social media accounts that were opened using the OWASP registered trademark word or logo need to be closed to avoid trademark infringement. The OWASP® Word Mark and OWASP & Design™ Logo are registered or unregistered service marks of OWASP Foundation, Inc. in the United States and other countries. All rights reserved. Unauthorized use is strictly prohibited.
 
If you wish to appeal the closing. You will need to submit the request via email with the justification to the OWASP Chapter Committee Leaders. The OWASP Chapter Committee meets on the second Tuesday of every month via zoom. The next meeting is November 8th at 1 pm EST.

Need additional assistance or have questions, Contact Us.
Become an OWASP Member.